Add authAPjs
